pyproject.toml
[project]
name = "koval"
version = "0.1.0"
description = "The Blacksmith for your Codebase."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm",
    "textual",
    "gitpython",
    "typer",
    "pyyaml"
]

[project.scripts]
koval = "main:app"
